Eval ai (#42)
* Adding EvalAI code

Adding the EvalAI code and documents to the EvalAI
folder. The contents of EvalAI can be zipped and used
to create an EvalAI challenge. The zipping wil be
run automatically by run.sh (this script needs to be
updated though). Also fixing some bugs in the testing
code.

* some modifications to the training scripts

* Added readme for eval ai training

Also small change to training scripts

* editing evalai descriptions and scripts

* Adding newlines to the html text